// src/api/__generated__/mining/model/cursorPageResponseRewardDetailDTO.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Mining API
 * 挖矿服务接口
 * OpenAPI spec version: v1.0
 */
import type { RewardDetailDTO } from './rewardDetailDTO';

/**
 * 游标分页响应
 */
export interface CursorPageResponseRewardDetailDTO {
  /** 每页数量 */
  pageSize?: string;
  /** 下一页游标，-1 表示没有更多数据 */
  mark?: string;
  /** 是否还有更多数据 */
  hasMore?: boolean;
  /** 分页数据 */
  list?: RewardDetailDTO[];
}