// src/api/__generated__/mining/model/listRewardDetailsType.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Mining API
 * 挖矿服务接口
 * OpenAPI spec version: v1.0
 */

export type ListRewardDetailsType =
  (typeof ListRewardDetailsType)[keyof typeof ListRewardDetailsType];

export const ListRewardDetailsType = {
  ALL: 'ALL',
  MINING: 'MINING',
  INVITE: 'INVITE',
} as const;