// src/api/__generated__/mining/model/rewardDetailDTO.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Mining API
 * 挖矿服务接口
 * OpenAPI spec version: v1.0
 */
import type { RewardDetailDTOAuditStatus } from './rewardDetailDTOAuditStatus';
import type { RewardDetailDTOType } from './rewardDetailDTOType';

/**
 * 奖励明细响应
 */
export interface RewardDetailDTO {
  /** 奖励时间 */
  rewardTime?: string;
  /** 奖励类型 */
  type?: RewardDetailDTOType;
  /** 奖励周期开始时间 */
  rewardPeriodStart?: string;
  /** 奖励周期结束时间 */
  rewardPeriodEnd?: string;
  /**
   * 来源用户ID（邀请奖励时有效）
   * @nullable
   */
  sourceUser?: number | null;
  /**
   * 来源用户名称（邀请奖励时有效）
   * @nullable
   */
  sourceUserName?: string | null;
  /**
   * STORY数量
   * @minimum 0
   */
  storyAmount?: number;
  /** PENDING/APPROVED/REJECTED */
  auditStatus?: RewardDetailDTOAuditStatus;
}