// src/api/__generated__/mining/model/rewardDetailDTOType.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Mining API
 * 挖矿服务接口
 * OpenAPI spec version: v1.0
 */

/**
 * 奖励类型
 */
export type RewardDetailDTOType =
  (typeof RewardDetailDTOType)[keyof typeof RewardDetailDTOType];

export const RewardDetailDTOType = {
  MINING: 'MINING',
  INVITE: 'INVITE',
} as const;